Refactor: Move Main & Splash activity to activites package
